pyproject.toml
[project]
name = "alpha-trader"
version = "0.0.1a1"
description = "AI-powered covered call & CSP advisor for options trading"
requires-python = ">=3.10"
dependencies = [
    "yfinance>=0.2.0",
    "pyyaml>=6.0",
    "flask>=3.0",
    "requests>=2.28",
    "scikit-learn>=1.3",
    "joblib>=1.3",
]

[project.scripts]
alpha-trader = "src.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
